refactor: split sim scene schema into modules
scene.py had grown past the 1000-line module limit. Split along
the natural seam: scene_schema.py holds the key inventories, the
information-boundary classification, TRUTH_KEYS, and the
import-time completeness assertion; scene_checks.py holds the
per-block field validators and primitive type checks; scene.py
remains the public entry point (load/save/validate,
build_nav_params, catalog helpers) and re-exports the boundary
names, so every existing import path keeps working. The ring-key
comment now describes current state (the rings list and its keys
remain valid until a ring-system block replaces them).
